pyproject.toml
[project]
name = "papertrade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair<5",
    "black>=26.1.0",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
    "sqlmodel>=0.0.31",
    "streamlit>=1.19.0",
    "typer>=0.21.1",
    "yfinance>=1.1.0",
]

[tool.setuptools]
packages = ["app"]

